Working v1.5 implemented

+++ src/deserializer/ProcessElement.java
@@ -2,7 +2,7 @@ package deserializer;
 
 import java.util.ArrayList;
 
-public record ProcessElement(int proc_id, int ord_dish, String proc_started, String proc_ended,
+public record ProcessElement(long proc_id, long ord_dish, String proc_started, String proc_ended,
                              boolean proc_active, ArrayList<OperationID> proc_operations) {
 
 }
